refactor: Standardize components in TransactionDetailModal
