+++ src/components/Modals/TransactionDetailModal/Data/Data.types.ts
@@ -1,0 +1,4 @@
+export type Props = {
+  label: string
+  children: React.ReactNode
+}
