+++ src/components/Modals/TransactionDetailModal/Data/index.ts
@@ -1,0 +1,2 @@
+import Data from './Data'
+export default Data
