+++ src/components/Modals/TransactionDetailModal/ExplorerLink/ExplorerLink.types.tsx
@@ -1,0 +1,7 @@
+import { ChainId, Network } from '@dcl/schemas'
+
+export type Props = {
+  network?: Network
+  chainId?: ChainId
+  txHash: string
+}
