+++ src/components/Modals/TransactionDetailModal/ExplorerLink/index.ts
@@ -1,0 +1,2 @@
+import ExplorerLink from './ExplorerLink'
+export default ExplorerLink
