+++ src/components/Modals/TransactionDetailModal/WithdrawalDataComponent/WithdrawalDataComponent.types.ts
@@ -1,0 +1,8 @@
+import { Transaction, Withdrawal } from '../../../../modules/mana/types'
+
+export type Props = {
+  withdrawals: Withdrawal[]
+  data: Transaction['data']
+}
+
+export type MapStateProps = Pick<Props, 'withdrawals'>
