+++ src/components/Modals/TransactionDetailModal/WithdrawalDataComponent/index.ts
@@ -1,0 +1,2 @@
+import WithdrawalDataComponent from './WithdrawalDataComponent.container'
+export default WithdrawalDataComponent
